Refaktor: Flyttet metoder i RelasjonerFunksjoner til Familierelasjon.
