+++ src/main/kotlin/no/nav/medlemskap/domene/Personhistorikk.kt
@@ -1,5 +1,7 @@
 package no.nav.medlemskap.domene
+import no.bekk.bekkopen.person.FodselsnummerValidator
 import no.nav.medlemskap.common.statsborgerskapUavklartForRegel
+import no.nav.medlemskap.domene.Fødselsnummer.Companion.hentBursdagsAar
 import no.nav.medlemskap.regler.common.RegelId
 import java.time.LocalDate
 import java.time.LocalDateTime
@@ -130,7 +132,33 @@ data class Familierelasjon(
     val relatertPersonsRolle: Familierelasjonsrolle,
     val minRolleForPerson: Familierelasjonsrolle?,
     val folkeregistermetadata: Folkeregistermetadata?
-)
+) {
+    companion object {
+        fun String.erBarnUnder25Aar(førsteDatoForYtelse: LocalDate) =
+            (førsteDatoForYtelse.year - this.hentBursdagsAar().toInt()) <= 25
+
+        fun hentFnrTilEktefelle(personHistorikkFraPdl: Personhistorikk?): String? {
+            val fnrTilEktefelle =
+                personHistorikkFraPdl?.sivilstand
+                    ?.filter { it.relatertVedSivilstand != null }
+                    ?.filter { FodselsnummerValidator.isValid(it.relatertVedSivilstand) }
+                    ?.filter { it.type == Sivilstandstype.GIFT || it.type == Sivilstandstype.REGISTRERT_PARTNER }
+                    ?.map { it.relatertVedSivilstand }
+                    ?.lastOrNull()
+            return fnrTilEktefelle
+        }
+
+        fun hentFnrTilBarn(familierelasjoner: List<Familierelasjon>, førsteDatoForYtelse: LocalDate): List<String> {
+            return familierelasjoner
+                .filter { FodselsnummerValidator.isValid(it.relatertPersonsIdent) }
+                .filter { it.erBarn() }
+                .filter { it.relatertPersonsIdent.erBarnUnder25Aar(førsteDatoForYtelse) }
+                .map { it.relatertPersonsIdent }
+        }
+
+        fun Familierelasjon.erBarn() = this.relatertPersonsRolle == Familierelasjonsrolle.BARN
+    }
+}
 
 enum class Familierelasjonsrolle {
     BARN, MOR, FAR, MEDMOR
