+++ src/test/kotlin/no/nav/medlemskap/regler/funksjoner/FamilierelasjonTest.kt
@@ -1,12 +1,13 @@
 package no.nav.medlemskap.clients.pdl.mapper
 
 import no.nav.medlemskap.domene.*
-import no.nav.medlemskap.regler.funksjoner.RelasjonFunksjoner
+import no.nav.medlemskap.domene.Familierelasjon.Companion.hentFnrTilBarn
+import no.nav.medlemskap.domene.Familierelasjon.Companion.hentFnrTilEktefelle
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import java.time.LocalDate
 
-class RelasjonFunksjonerTest {
+class FamilierelasjonTest {
 
     @Test
     fun `hent fnr til ektefelle`() {
@@ -15,7 +16,7 @@ class RelasjonFunksjonerTest {
                 lagSivilstand(relatertVedSivilstand = "13128619857")
             )
         )
-        val fnrTilEktefelle = RelasjonFunksjoner.hentFnrTilEktefelle(personhistorikk)
+        val fnrTilEktefelle = hentFnrTilEktefelle(personhistorikk)
         assertEquals("13128619857", fnrTilEktefelle)
     }
 
@@ -28,7 +29,7 @@ class RelasjonFunksjonerTest {
         )
 
         val førsteDatoForYtelse = LocalDate.of(2020, 12, 12)
-        val fnrTilBarn = RelasjonFunksjoner.hentFnrTilBarn(familierelasjoner, førsteDatoForYtelse)
+        val fnrTilBarn = hentFnrTilBarn(familierelasjoner, førsteDatoForYtelse)
         val forventetSvar = listOf("25079528660", "09069534888")
         assertEquals(2, fnrTilBarn.size)
         assertEquals(forventetSvar, fnrTilBarn)
